Change int stats to uint64
Other int stats changed from int64 to int, where it made sense.

+++ client/nginx_client.go
@@ -22,9 +22,9 @@ type versions []int
 
 // UpstreamServer lets you configure HTTP upstreams
 type UpstreamServer struct {
-	ID          int64  `json:"id,omitempty"`
+	ID          int    `json:"id,omitempty"`
 	Server      string `json:"server"`
-	MaxFails    int64  `json:"max_fails"`
+	MaxFails    int    `json:"max_fails"`
 	FailTimeout string `json:"fail_timeout,omitempty"`
 	SlowStart   string `json:"slow_start,omitempty"`
 }
@@ -68,23 +68,23 @@ type Stats struct {
 
 // Connections metrics
 type Connections struct {
-	Accepted int64
-	Dropped  int64
-	Active   int64
-	Idle     int64
+	Accepted uint64
+	Dropped  uint64
+	Active   uint64
+	Idle     uint64
 }
 
 // HTTPRequests metrics
 type HTTPRequests struct {
-	Total   int64
-	Current int64
+	Total   uint64
+	Current uint64
 }
 
 // SSL metrics
 type SSL struct {
-	Handshakes       int64
-	HandshakesFailed int64 `json:"handshakes_failed"`
-	SessionReuses    int64 `json:"session_reuses"`
+	Handshakes       uint64
+	HandshakesFailed uint64 `json:"handshakes_failed"`
+	SessionReuses    uint64 `json:"session_reuses"`
 }
 
 // ServerZones collection
@@ -92,21 +92,21 @@ type ServerZones map[string]ServerZone
 
 // ServerZone metrics
 type ServerZone struct {
-	Processing int64
-	Requests   int64
+	Processing uint64
+	Requests   uint64
 	Responses  Responses
-	Discarded  int64
-	Received   int64
-	Sent       int64
+	Discarded  uint64
+	Received   uint64
+	Sent       uint64
 }
 
 // Responses metrics
 type Responses struct {
-	Responses1xx int64 `json:"1xx"`
-	Responses2xx int64 `json:"2xx"`
-	Responses3xx int64 `json:"3xx"`
-	Responses4xx int64 `json:"4xx"`
-	Responses5xx int64 `json:"5xx"`
+	Responses1xx uint64 `json:"1xx"`
+	Responses2xx uint64 `json:"2xx"`
+	Responses3xx uint64 `json:"3xx"`
+	Responses4xx uint64 `json:"4xx"`
+	Responses5xx uint64 `json:"5xx"`
 }
 
 // Upstreams collection
@@ -125,7 +125,7 @@ type Upstream struct {
 type Queue struct {
 	Size      int
 	MaxSize   int `json:"max_size"`
-	Overflows int64
+	Overflows uint64
 }
 
 // Peer metrics
@@ -137,27 +137,27 @@ type Peer struct {
 	Backup       bool
 	Weight       int
 	State        string
-	Active       int64
+	Active       uint64
 	MaxConns     int `json:"max_conns"`
-	Requests     int64
+	Requests     uint64
 	Responses    Responses
-	Sent         int64
-	Received     int64
-	Fails        int64
-	Unavail      int64
+	Sent         uint64
+	Received     uint64
+	Fails        uint64
+	Unavail      uint64
 	HealthChecks HealthChecks
-	Downtime     int64
+	Downtime     int
 	Downstart    string
 	Selected     string
-	HeaderTime   int64 `json:"header_time"`
-	ResponseTime int64 `json:"response_time"`
+	HeaderTime   int `json:"header_time"`
+	ResponseTime int `json:"response_time"`
 }
 
 // HealthChecks metrics
 type HealthChecks struct {
-	Checks     int64
-	Fails      int64
-	Unhealthy  int64
+	Checks     uint64
+	Fails      uint64
+	Unhealthy  uint64
 	LastPassed bool `json:"last_passed"`
 }
 
@@ -354,7 +354,7 @@ func determineUpdates(updatedServers []UpstreamServer, nginxServers []UpstreamSe
 	return
 }
 
-func (client *NginxClient) getIDOfHTTPServer(upstream string, name string) (int64, error) {
+func (client *NginxClient) getIDOfHTTPServer(upstream string, name string) (int, error) {
 	servers, err := client.GetHTTPServers(upstream)
 	if err != nil {
 		return -1, fmt.Errorf("error getting id of server %v of upstream %v: %v", name, upstream, err)
